src/requirements.txt
streamlit
pandas
scikit-learn
fuzzywuzzy
python-Levenshtein
numpy
requests
python-dotenv
